node_modules: add default source filtering and add opt-out
Very often we can install all the node packages without any of the stuff
in the source tree. The only relevant pieces are the `package-lock.json`
and the `package.json`. By default we will now throw away all files
except the `package.json`.

This also adds a `filterSource` boolean that allows opt-out of the
source filtering - if that is deisred. We also allow the users to
provide another filter function (name `sourceFilter`) to customize the
filtering.